feat: admin 예약관리 상세에서 이미지 추가, 삭제 기능 구현
API에서 id를 pathvariable로 받도록 변경하고, 이미지나 step을 추가하는 post API에서 response에 id값을 담아주도록 변경했습니다.
