--- admin/.env
@@ -1,2 +1,0 @@
-REACT_APP_REDIRECT_URI=http://localhost:3000
-REACT_APP_SERVER_URL=http://localhost:8080
