+++ admin/src/api/Settings.js
@@ -22,4 +22,3 @@ const axios = Axios.create({
 // );
 
 export default axios;
-/* github actions test */
