+++ server/src/main/java/com/genesisairport/reservation/common/util/GlobalExceptionHandler.java
@@ -25,7 +25,7 @@ public class GlobalExceptionHandler extends ResponseEntityExceptionHandler {
     }
 
     private ResponseEntity<Object> handleExceptionInternal(Exception e, ResponseCode responseCode,
-                                                           HttpHeaders headers, WebRequest webRequest){
+                                                           HttpHeaders headers, WebRequest webRequest) {
 
         DataResponseDto<Object> body = DataResponseDto.of(null, responseCode.getMessage(e));
