+++ server/src/main/java/com/genesisairport/reservation/response/AdminResponse.java
@@ -27,4 +27,11 @@ public class AdminResponse {
         private String date;
         private List<String> availableTime;
     }
+
+    @Getter
+    @Builder
+    public static class UploadImage {
+        private Long imageId;
+        private String imageUrl;
+    }
 }
